+++ MovappUITests/XCUIApplication+arguments.swift
@@ -13,4 +13,23 @@ extension XCUIApplication {
         // The key must be se same as user default key
         launchArguments += ["-onboarding.complete", seen ? "true" : "false"]
     }
+    
+    func setLanguageTo(_ languageKey: String) {
+        launchArguments += ["-AppleLanguages", "(\(languageKey))"]
+        
+        let localKey = languageKey.replacingOccurrences(of: "-", with: "_")
+        launchArguments += ["-AppleLocale", "\"\(localKey)\""]
+    }
+    
+    func setLanguageToUk() {
+        setLanguageTo("uk")
+    }
+    
+    func setLanguageToEn() {
+        setLanguageTo("en-US")
+    }
+    
+    func setLanguageToCs() {
+        setLanguageTo("uk")
+    }
 }