+++ Shared/Stores/TeamDataStore.swift
@@ -5,8 +5,7 @@
 //  Created by Martin Kluska on 02.05.2022.
 //
 
-import Foundation
-import UIKit
+import SwiftUI
 
 class TeamDataStore: ObservableObject {